assignment 2

+++ src/album/album.service.js
@@ -1,40 +1,60 @@
 const albumModel = require("./album.model");
 const songModel = require("../song/song.model");
-const statusCode = require("../errorCodes");
+const statusCodes = require("../configs/errorCodes.config");
+const errorMessages = require("../errorMessages");
+const moment = require("moment");
 
-class ItemService {
+class albumService {
   async getAlbums() {
     return await albumModel.find();
   }
 
-  async addAlbum(albumObject) {
-    return await new albumModel({
-      name: albumObject.name,
-      description: albumObject.description,
-      showNbTracks: albumObject.showNbTracks,
+  async addAlbum(reqBody, userId) {
+    await new albumModel({
+      name: reqBody.name,
+      description: reqBody.description,
+      showNbTracks: reqBody.showNbTracks,
+      createdBy: userId,
     }).save();
   }
 
   //A check was added to notify the user if he provides a wrong albumId
   async getAlbumById(albumId) {
-    const foundAlbum = await albumModel.findOne({ _id: albumId });
-    if (!foundAlbum) {
-      const error = new Error("Album not found");
-      error.statusCode = statusCode.notFound;
+    const album = await albumModel.findOne({ _id: albumId });
+    if (!album) {
+      const error = new Error(errorMessages.albumNotFound);
+      error.statusCode = statusCodes.notFound;
       throw error;
     } else {
-      return foundAlbum;
+      return album;
     }
   }
-
-  async updateAlbum(albumId, albumObject) {
-    await albumModel.updateOne({ _id: albumId }, { $set: albumObject });
+  //the getAlbumById function was used to check that the user entered a valid albumId
+  async updateAlbum(albumId, reqBody, userId) {
+    await this.getAlbumById(albumId);
+    await albumModel.updateOne(
+      { _id: albumId },
+      { $set: reqBody, updatedBy: userId }
+    );
   }
-
-  //All songs added to the album to delete will be deleted
+  //the getAlbumById function was used to check that the user entered a valid albumId
   async deleteAlbum(albumId) {
-    await songModel.deleteMany({ albumId: albumId });
-    await albumModel.deleteOne({ _id: albumId });
+    await this.getAlbumById(albumId);
+    const relatedSong = await songModel.findOne({ albumId: albumId });
+    if (relatedSong) {
+      const error = new Error(errorMessages.cannotDeleteAlbum);
+      error.statusCode = statusCodes.conflict;
+      throw error;
+    } else {
+      await albumModel.deleteOne({ _id: albumId });
+    }
+  }
+  //this function was used in the addSong function in the song service
+  async updateLastSongAddedAt(albumId, userId) {
+    await albumModel.updateOne(
+      { _id: albumId },
+      { $set: { lastSongAddedAt: moment(), updatedBy: userId } }
+    );
   }
 }
-module.exports = ItemService;
+module.exports = albumService;
